Move utils outside graphite package
