+++ graphite/client.go
@@ -148,7 +148,7 @@ func (c *Client) queryToTargets(query *prompb.Query, ctx context.Context) ([]str
 
 	// Prepare the url to fetch
 	queryStr := c.prefix + name + ".**"
-	expandUrl, err := prepareUrl(c.graphite_web, expandEndpoint, map[string]string{"format": "json", "leavesOnly": "1", "query": queryStr})
+	expandUrl, err := prepareURL(c.graphite_web, expandEndpoint, map[string]string{"format": "json", "leavesOnly": "1", "query": queryStr})
 	if err != nil {
 		log.With("graphite_web", c.graphite_web).With("path", expandEndpoint).With("err", err).Warnln("Error preparing URL")
 		return nil, err
@@ -156,7 +156,7 @@ func (c *Client) queryToTargets(query *prompb.Query, ctx context.Context) ([]str
 
 	// Get the list of targets
 	expandResponse := ExpandResponse{}
-	body, err := fetchUrl(expandUrl, ctx)
+	body, err := fetchURL(expandUrl, ctx)
 	if err != nil {
 		log.With("url", expandUrl).With("err", err).Warnln("Error fetching URL")
 		return nil, err
@@ -210,14 +210,14 @@ func (c *Client) filterTargets(query *prompb.Query, targets []string) ([]string,
 }
 
 func (c *Client) targetToTimeseries(target string, from string, until string, ctx context.Context) (*prompb.TimeSeries, error) {
-	renderUrl, err := prepareUrl(c.graphite_web, renderEndpoint, map[string]string{"format": "json", "from": from, "until": until, "target": target})
+	renderUrl, err := prepareURL(c.graphite_web, renderEndpoint, map[string]string{"format": "json", "from": from, "until": until, "target": target})
 	if err != nil {
 		log.With("graphite_web", c.graphite_web).With("path", renderEndpoint).With("err", err).Warnln("Error preparing URL")
 		return nil, err
 	}
 
 	renderResponses := make([]RenderResponse, 0)
-	body, err := fetchUrl(renderUrl, ctx)
+	body, err := fetchURL(renderUrl, ctx)
 	if err != nil {
 		log.With("url", renderUrl).With("err", err).Warnln("Error fetching URL")
 		return nil, err
