+++ graphite/rules.go
@@ -25,7 +25,7 @@ import (
 	"github.com/prometheus/prometheus/prompb"
 
 	"github.com/criteo/graphite-remote-adapter/graphite/config"
-	"github.com/criteo/graphite-remote-adapter/graphite/utils"
+	"github.com/criteo/graphite-remote-adapter/utils"
 
 	"github.com/patrickmn/go-cache"
 )
